Move file handling objects into fs.rs
